feat: implement auth message creation
