+++ crates/core/src/rlp/encode.rs
@@ -5,6 +5,14 @@ use tinyvec::ArrayVec;
 
 use super::constants::RLP_NULL;
 
+/// Function for encoding a value to RLP.
+/// For encoding the value into a buffer directly, use [`RLPEncode::encode`].
+pub fn encode<T: RLPEncode>(value: T) -> Vec<u8> {
+    let mut buf = Vec::new();
+    value.encode(&mut buf);
+    buf
+}
+
 pub trait RLPEncode {
     fn encode(&self, buf: &mut dyn BufMut);
